# openapi.yaml
openapi: 3.0.3
info:
  title: Kabanchik Service API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/logout:
    post:
      summary: Logout
      responses:
        '204':
          description: No Content
  /me:
    get:
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /services:
    get:
      summary: List services
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: min_price
          schema: { type: integer, format: int64 }
        - in: query
          name: max_price
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
    post:
      summary: Create service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
  /services/{id}:
    get:
      summary: Get service
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
    patch:
      summary: Update service
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '200':
          description: OK
    delete:
      summary: Delete service
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204':
          description: No Content
  /orders:
    get:
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      summary: Get order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    patch:
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '204':
          description: No Content
  /orders/{id}/message:
    post:
      summary: Add order message
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderMessageRequest'
      responses:
        '204':
          description: No Content

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string }
        password: { type: string }
        role: { type: string, enum: [client, provider] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        city: { type: string }
    CreateServiceRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        city: { type: string }
        price: { type: integer, format: int64 }
    CreateOrderRequest:
      type: object
      required: [service_id]
      properties:
        service_id: { type: string }
        details: { type: string }
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [new, accepted, done, canceled] }
    OrderMessageRequest:
      type: object
      required: [text]
      properties:
        text: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        role: { type: string }
        name: { type: string }
        phone: { type: string }
        city: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
    Service:
      type: object
      properties:
        id: { type: string }
        provider_id: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        city: { type: string }
        price: { type: integer, format: int64 }
        created_at: { type: string }
        updated_at: { type: string }
    Order:
      type: object
      properties:
        id: { type: string }
        service_id: { type: string }
        client_id: { type: string }
        provider_id: { type: string }
        status: { type: string }
        details: { type: string }
        created_at: { type: string }
        updated_at: { type: string }